StoneBoard: remove clear() method. StoneBoard: constructor calls startNewGame().
No more virtual calls to clear() (which was only called in
startNewGame()). This would cause GroupBoard and PatternBoard to clear
their data, which is pretty much pointless because it will be cleared
when absorb() and update() are called by hand later. I don't think there
is a case where startNewGame() is called and update()/absorb() are not
called by hand before that data is used.

No need to call startNewGame() after constructing a StoneBoard now, since
it is called automatically. Which is what you would expect it to do.